🚚 Rename Spotify authentication file
